move wallet state to state lib and wallet classes to its own lib folder
