+++ components/account/OnboardingModal.tsx
@@ -1,9 +1,9 @@
 import { Dialog } from "@headlessui/react";
 import { useOnboarding } from "lib/state/onboarding";
-import { BaseDotsamaWallet } from "lib/stores/wallets/base-dotsama-wallet";
-import { PolkadotjsWallet } from "lib/stores/wallets/polkadotjs-wallet";
-import { SubWallet } from "lib/stores/wallets/subwallet";
-import { TalismanWallet } from "lib/stores/wallets/talisman-wallet";
+import { BaseDotsamaWallet } from "lib/wallets/base-dotsama-wallet";
+import { PolkadotjsWallet } from "lib/wallets/polkadotjs-wallet";
+import { SubWallet } from "lib/wallets/subwallet";
+import { TalismanWallet } from "lib/wallets/talisman-wallet";
 import { range } from "lodash-es";
 import { observer } from "mobx-react";
 import Image from "next/image";
