+++ components/outcomes/ScalarReportBox.tsx
@@ -11,7 +11,7 @@ import { ZTG } from "lib/constants";
 import { useSdkv2 } from "lib/hooks/useSdkv2";
 import { useNotifications } from "lib/state/notifications";
 import { useStore } from "lib/stores/Store";
-import { useWallet } from "lib/stores/wallets";
+import { useWallet } from "lib/state/wallet";
 import { getCurrentPrediction } from "lib/util/assets";
 import { extrinsicCallback, signAndSend } from "lib/util/tx";
 import { observer } from "mobx-react";
