+++ lib/wallets/errors/AuthError.tsx
