+++ lib/wallets/errors/BaseWalletError.tsx
